# perf: cache the creation of the algorithm